:recycle: (refactor) Put "project" slash command logic in dedicated …
